pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slm-api-server"
version = "0.1.0"
description = "Flask API wrapper around llama-cpp-python with SSE and local GGUF loader"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "llama-cpp-python>=0.2.90",
  "huggingface_hub>=0.23",
  "python-dotenv>=1.0.1",
  "gunicorn>=21.2.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "httpx>=0.27.0", "ruff>=0.5.0"]

[tool.ruff]
line-length = 100